меняет рендер и remove для работы с компонентами
