+++ src/components/abstract-component.js
@@ -1,9 +1,9 @@
-import {createElement} from "./utils";
+import {createElement} from "../utils/render";
 
 class AbstractComponent {
   constructor() {
     if (new.target === AbstractComponent) {
-      throw  new Error(`Can't insnantiate AbstractComponent, only concrete one.`);
+      throw new Error(`Can't insnantiate AbstractComponent, only concrete one.`);
     }
 
     this._element = null;
