+++ src/main.js
@@ -12,22 +12,22 @@ import {generateTasks} from "./mock/task";
 
 import {OptionTasks, Place} from "./components/consts";
 
-import {render} from "./components/utils";
+import {render, replace, remove} from "./utils/render";
 
 let openTask;
 
 const renderTask = (taskListElement, task) => {
   const replaceTaskToEdit = () => {
     if (openTask) {
-      taskListElement.replaceChild(openTask[0], openTask[1]);
+      replace(taskListElement, openTask[0], openTask[1]);
       document.removeEventListener(`keydown`, onEscKeyDown);
     }
     openTask = [taskComponent.getElement(), taskEditComponent.getElement()];
-    taskListElement.replaceChild(taskEditComponent.getElement(), taskComponent.getElement());
+    replace(taskListElement, taskEditComponent.getElement(), taskComponent.getElement());
   };
 
   const replaceEditToTask = () => {
-    taskListElement.replaceChild(taskComponent.getElement(), taskEditComponent.getElement());
+    replace(taskListElement, taskComponent.getElement(), taskEditComponent.getElement());
     openTask = null;
   };
 
@@ -56,18 +56,18 @@ const renderTask = (taskListElement, task) => {
     document.removeEventListener(`keydown`, onEscKeyDown);
   });
 
-  render(taskListElement, taskComponent.getElement(), Place.BEFOREEND);
+  render(taskListElement, taskComponent, Place.BEFOREEND);
 };
 
 const renderBoard = (boardComponent, tasks) => {
   const isAllTasksArchived = tasks.every((task) => task.isArchive);
 
   if (isAllTasksArchived) {
-    render(boardComponent.getElement(), new NoTasksComponent().getElement(), Place.BEFOREEND);
+    render(boardComponent.getElement(), new NoTasksComponent(), Place.BEFOREEND);
     return;
   }
 
-  render(boardComponent.getElement(), new TasksComponent().getElement(), Place.BEFOREEND);
+  render(boardComponent.getElement(), new TasksComponent(), Place.BEFOREEND);
 
   const taskListElement = boardComponent.getElement().querySelector(`.board__tasks`);
 
@@ -77,15 +77,14 @@ const renderBoard = (boardComponent, tasks) => {
     });
 
   const loadMoreButtonComponent = new MoreButtonComponent();
-  render(boardComponent.getElement(), loadMoreButtonComponent.getElement(), Place.BEFOREEND);
+  render(boardComponent.getElement(), loadMoreButtonComponent, Place.BEFOREEND);
 
   const onMoreView = () => {
     tasksCopy.splice(0, OptionTasks.MORE_SHOW)
       .forEach((task) => renderTask(taskListElement, task));
 
     if (tasksCopy.length === 0) {
-      loadMoreButtonComponent.getElement().remove();
-      loadMoreButtonComponent.removeElement();
+      remove(loadMoreButtonComponent);
     }
   };
 
@@ -99,11 +98,11 @@ const filters = generateFilters();
 const tasks = generateTasks(OptionTasks.COUNT);
 const tasksCopy = tasks.slice(OptionTasks.START_SHOW);
 
-render(siteHeaderElement, new SiteMenuComponent().getElement(), Place.BEFOREEND);
-render(siteMainElement, new FilterComponent(filters).getElement(), Place.BEFOREEND);
+render(siteHeaderElement, new SiteMenuComponent(), Place.BEFOREEND);
+render(siteMainElement, new FilterComponent(filters), Place.BEFOREEND);
 
 const boardComponent = new BoardComponent();
 
-render(siteMainElement, boardComponent.getElement(), Place.BEFOREEND);
+render(siteMainElement, boardComponent, Place.BEFOREEND);
 
 renderBoard(boardComponent, tasks);
